refactor: split functions/tests according to usage
